requirements.txt
Pillow
pil-utils
requests
numpy
opencv-python-headless
meme-generator